run tests in typescript (#85)
* run tests in typescript

* oops, fix test

* fix linting

* real fixes

* use more yarn

* fix lint

* try this

* revert coverage reporting

* ok whatever

* might aswell change these too

* ugh, fix this

* I cba anymore
