+++ scripts/.travis-deploy.sh
@@ -5,7 +5,7 @@ set -e
 
 if [ "$TRAVIS_BRANCH" != "master" -o -n "$TRAVIS_TAG" -o "$TRAVIS_PULL_REQUEST" != "false" ]; then
   echo -e "Not building for a non master branch push - building without deploying."
-  npm run docs
+  yarn run docs
   exit 0
 fi
 
@@ -17,7 +17,7 @@ SHA=$(git rev-parse --verify HEAD)
 TARGET_BRANCH="gh-pages"
 git clone $REPO dist -b $TARGET_BRANCH
 
-npm run docs
+yarn run docs
 
 rsync -vau docs/ dist/
