+++ test/writeFileAtomic.ts
@@ -1,7 +1,7 @@
-const ava = require('ava');
-const { promises: fs } = require('fs');
-const { tempFile, tempFileLoc } = require('./lib');
-const nextra = require('../dist');
+import ava from 'ava';
+import { promises as fs } from 'fs';
+import { tempFile, tempFileLoc } from './lib';
+import * as nextra from '../dist';
 
 ava('Existing', async test => {
 	test.plan(2);
